# architect-career-api-sdk/openapi/acos-api.yaml
openapi: 3.1.0
info:
  title: Architect Career Operating System API
  description: Platform API documentation for acos-platform
  contact:
    name: ACOS Engineering
    email: platform@acos.local
  license:
    name: Proprietary
  version: development
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
security:
  - bearer-jwt: []
...
paths:
...
  /api/v1/career/companies/{companyId}:
    get:
      tags:
        - Career Companies
      summary: Get company
      description: Returns a company owned by the authenticated user.
      operationId: get_9
      parameters:
        - name: companyId
          in: path
          description: Company identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerCompanyApiResponse'
        '401':
          description: JWT authentication required for company APIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '404':
          description: Company not found for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
      security:
        - bearer-jwt: []
    put:
      tags:
        - Career Companies
      summary: Update company
      description: Updates a company owned by the authenticated user.
      operationId: update_10
      parameters:
        - name: companyId
          in: path
          description: Company identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerCompanyApiResponse'
        '400':
          description: Company request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '401':
          description: JWT authentication required for company APIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '404':
          description: Company not found for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
      security:
        - bearer-jwt: []
    delete:
      tags:
        - Career Companies
      summary: Delete company
      description: Deletes a company owned by the authenticated user. Companies referenced by job applications cannot be deleted.
      operationId: delete_10
      parameters:
        - name: companyId
          in: path
          description: Company identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerDeleteApiResponse'
        '401':
          description: JWT authentication required for company APIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '404':
          description: Company not found for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '422':
          description: Company is still referenced by job applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
      security:
        - bearer-jwt: []
...
  /api/v1/career/companies:
    get:
      tags:
        - Career Companies
      summary: List companies
      description: Lists all companies owned by the authenticated user.
      operationId: list_10
      responses:
        '200':
          description: Companies listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerCompanyListApiResponse'
        '401':
          description: JWT authentication required for company APIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
      security:
        - bearer-jwt: []
    post:
      tags:
        - Career Companies
      summary: Create company
      description: Creates a tracked company owned by the authenticated user. Duplicate names for the same owner are rejected.
      operationId: create_10
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerCompanyApiResponse'
        '400':
          description: Company request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '401':
          description: JWT authentication required for company APIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
        '422':
          description: Business rule violation (for example duplicate company name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerErrorApiResponse'
      security:
        - bearer-jwt: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDetail'
...
    FieldErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue: {}
...
    CareerErrorApiResponse:
      type: object
      description: Career error response envelope
      properties:
        success:
          type: boolean
          description: Whether the request succeeded
          example: false
        data:
          type: 'null'
          description: Success payload
          writeOnly: true
        error:
          $ref: '#/components/schemas/ApiError'
          description: Error payload
        correlationId:
          type: string
          description: Request correlation identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2026-08-04T06:00:00Z'
    CompanyRequest:
      type: object
      description: Payload used to create or update a company
      properties:
        name:
          type: string
          description: Company name
          example: Acme Corp
          maxLength: 200
          minLength: 0
        website:
          type:
            - string
            - 'null'
          description: Optional website
          example: https://acme.example.com
          maxLength: 500
          minLength: 0
        industry:
          type:
            - string
            - 'null'
          description: Optional industry
          example: Technology
          maxLength: 100
          minLength: 0
        location:
          type:
            - string
            - 'null'
          description: Optional location
          example: San Francisco, CA
          maxLength: 200
          minLength: 0
        notes:
          type:
            - string
            - 'null'
          description: Optional notes
          example: Enterprise customer focus
          maxLength: 2000
          minLength: 0
      required:
        - name
    CareerCompanyApiResponse:
      type: object
      description: Successful career company envelope
      properties:
        success:
          type: boolean
          description: Whether the request succeeded
          example: true
        data:
          $ref: '#/components/schemas/CompanyResponse'
          description: Company payload
        error:
          type: 'null'
          $ref: '#/components/schemas/ApiError'
          description: Error payload
        correlationId:
          type: string
          description: Request correlation identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2026-08-04T06:00:00Z'
    CompanyResponse:
      type: object
      description: Tracked company
      properties:
        id:
          type: string
          format: uuid
          description: Company identifier
        name:
          type: string
          description: Company name
        website:
          type:
            - string
            - 'null'
          description: Optional website
        industry:
          type:
            - string
            - 'null'
          description: Optional industry
        location:
          type:
            - string
            - 'null'
          description: Optional location
        notes:
          type:
            - string
            - 'null'
          description: Optional notes
        archived:
          type: boolean
          description: Whether the company is archived
        archivedAt:
          type:
            - string
            - 'null'
          format: date-time
          description: Optional archival timestamp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        version:
          type: integer
          format: int64
          description: Optimistic locking version
      required:
        - archived
        - createdAt
        - id
        - name
        - updatedAt
        - version
...
    CareerCompanyListApiResponse:
      type: object
      description: Successful career company list envelope
      properties:
        success:
          type: boolean
          description: Whether the request succeeded
          example: true
        data:
          type: array
          description: Companies payload
          items:
            $ref: '#/components/schemas/CompanyResponse'
        error:
          type: 'null'
          $ref: '#/components/schemas/ApiError'
          description: Error payload
        correlationId:
          type: string
          description: Request correlation identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2026-08-04T06:00:00Z'
...
    CareerDeleteApiResponse:
      type: object
      description: Successful career delete envelope
      properties:
        success:
          type: boolean
          description: Whether the request succeeded
          example: true
        data:
          type: 'null'
          description: Success payload
          writeOnly: true
        error:
          type: 'null'
          $ref: '#/components/schemas/ApiError'
          description: Error payload
        correlationId:
          type: string
          description: Request correlation identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2026-08-04T06:00:00Z'
  securitySchemes:
    bearer-jwt:
      type: http
      description: JWT access token issued by POST /api/v1/auth/login
      name: bearer-jwt
      scheme: bearer
      bearerFormat: JWT